Add KafkaConsumer component
This commit adds a Kafka Consumer that implements Stuart Sierra's component and provides high-level abstractions to work with in Clojure. To use this component, one must provide a valid mapping of topics and handlers like the example below:

```clj
(def consumer-topics
  {:topic-a
   {:topic   "TOPIC_A" ; this will be the name of the topic on Kafka
    :schema  SchemaA
    :handler (fn [message components] ; the handler signature takes the message and the components as parameters
               (do-some-stuff message components))}
   :topic-b
   {:topic   "TOPIC_B"
    :schema  SchemaB
    :handler (fn [message components]
               (do-some-other-cool-stuffs message components)}}})
```

And then add it to the system map like this:

```clj
(def my-system
  (component/system-map
    :config      (conf/new-config)

    :producer    (component/using
                   (kafka-producer/new-producer producer-topics)
                   [:config])

    :consumer    (component/using
                   (kafka-consumer/new-consumer consumer-topics) ; NOTE: the consumer depends on :config and :producer components
                   [:config :producer]))
```

Note that the consumer depends on both `config` and `producer` components. The dependency on producer is required because otherwise the consumer would not be able to produce deadletters in case of consumption failures. Also, for test environment the consumer listens to producer state to consume produced messages in memory.

Main features for now are:
- JSON deserialization
- Schema validation
- Components injection to handler (via `component/using` from Stuart Sierra)

In the future should add support for more data formats other than JSON.